// src/api/models/GenericTutorialStep.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import { TutorialStepBase } from './TutorialStepBase';

export type GenericTutorialStep = TutorialStepBase & {
  title: string;
  body: string;
  key?: undefined;
  imgUrl?: string;
};